helper_script/requirements.txt
psycopg2 == 2.9.3
mysql-connector-python == 8.0.25
sshtunnel == 0.4.0